requirements.txt
python-dotenv==1.0.0
dune-client==1.3.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0 